+++ packages/utils/src/types.ts
@@ -131,10 +131,12 @@ export type FieldId = {
 };
 
 /** Type describing a recursive structure of `FieldId`s for an object with a non-empty set of keys */
-export type IdSchema<T = any> = FieldId & {
-  /** The set of ids for fields in the recursive object structure */
-  [key in keyof T]?: IdSchema<T[key]>;
-};
+export type IdSchema<T = any> = T extends GenericObjectType
+  ? FieldId & {
+      /** The set of ids for fields in the recursive object structure */
+      [key in keyof T]?: IdSchema<T[key]>;
+    }
+  : FieldId;
 
 /** Type describing a name used for a field in the `PathSchema` */
 export type FieldPath = {
@@ -143,10 +145,16 @@ export type FieldPath = {
 };
 
 /** Type describing a recursive structure of `FieldPath`s for an object with a non-empty set of keys */
-export type PathSchema<T = any> = FieldPath & {
-  /** The set of names for fields in the recursive object structure */
-  [key in keyof T]?: PathSchema<T[key]>;
-};
+export type PathSchema<T = any> = T extends Array<infer U>
+  ? FieldPath & {
+      [i: number]: PathSchema<U>;
+    }
+  : T extends GenericObjectType
+  ? FieldPath & {
+      /** The set of names for fields in the recursive object structure */
+      [key in keyof T]?: PathSchema<T[key]>;
+    }
+  : FieldPath;
 
 /** The type for error produced by RJSF schema validation */
 export type RJSFValidationError = {
@@ -380,11 +388,11 @@ export interface FieldProps<T = any, S extends StrictRJSFSchema = RJSFSchema, F 
   /** A boolean value stating if the field should autofocus */
   autofocus?: boolean;
   /** A boolean value stating if the field is disabled */
-  disabled: boolean;
+  disabled?: boolean;
   /** A boolean value stating if the field is hiding its errors */
   hideError?: boolean;
   /** A boolean value stating if the field is read-only */
-  readonly: boolean;
+  readonly?: boolean;
   /** The required status of this field */
   required?: boolean;
   /** The unique name of the field, usually derived from the name of the property in the JSONSchema */
@@ -545,7 +553,7 @@ export type ArrayFieldTemplateItemType<
   /** The className string */
   className: string;
   /** A boolean value stating if the array item is disabled */
-  disabled: boolean;
+  disabled?: boolean;
   /** A boolean value stating whether new items can be added to the array */
   canAdd: boolean;
   /** A boolean value stating whether the array item can be copied, assumed false if missing */
@@ -571,7 +579,7 @@ export type ArrayFieldTemplateItemType<
   /** Returns a function that swaps the items at `index` with `newIndex` */
   onReorderClick: (index: number, newIndex: number) => (event?: any) => void;
   /** A boolean value stating if the array item is read-only */
-  readonly: boolean;
+  readonly?: boolean;
   /** A stable, unique key for the array item */
   key: string;
   /** The schema object for this array item */
@@ -631,9 +639,9 @@ export type ObjectFieldTemplatePropertyType = {
   /** A string representing the property name */
   name: string;
   /** A boolean value stating if the object property is disabled */
-  disabled: boolean;
+  disabled?: boolean;
   /** A boolean value stating if the property is read-only */
-  readonly: boolean;
+  readonly?: boolean;
   /** A boolean value stating if the property should be hidden */
   hidden: boolean;
 };